Implemented register page
A lot happened in the last 5 minutes of that video.

Alright so we've got a register.vue page with fields for taking email
and password. We bound these to variables using v-model, so we can
access and amend their values.

We created a register method that uses the authentication service we set
up previously. This calls the server with the credentials and gets
a response.

I had to install babel-runtime to avoid errors I was getting, which was
odd.